requirements.txt
apache-airflow
apache-airflow-providers-amazon
boto3
pandas
psycopg2-binary
SQLAlchemy
